shadow only some offset
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/branches/mipmap@6063 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CDepthBuffer.h
@@ -61,7 +61,7 @@ namespace video
 		virtual ~CStencilBuffer();
 
 		//! clears the zbuffer
-		virtual void clear(u32 value=0) _IRR_OVERRIDE_;
+		virtual void clear(u8 value=0) _IRR_OVERRIDE_;
 
 		//! sets the new size of the zbuffer
 		virtual void setSize(const core::dimension2d<u32>& size) _IRR_OVERRIDE_;
